Refactor ArrowOnLine component to utilize new utility functions for arrow calculation and polygon matching. Simplify onClick logic by integrating findMatchingPolygonIndex and reversePolygonPoints functions, enhancing code readability and maintainability.
